# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Food Delivery API
  version: 1.0.0
  description: REST API for a food delivery service.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/orders:
    post:
      tags:
        - Orders
      summary: Create an order
      description: Creates an order for the authenticated user using one of their saved delivery addresses.
      operationId: createOrder
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            example:
              address_id: 1
              items:
                - product_id: 10
                  quantity: 2
                - product_id: 15
                  quantity: 1

      responses:
        "201":
          description: Order successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "400":
          description: Invalid request body, address, product, or order data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Orders
      summary: List user orders
      description: Returns orders belonging to the authenticated user.
      operationId: listUserOrders
      security:
        - BearerAuth: []

      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Values less than or equal to zero are replaced with the default limit.

        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

      responses:
        "200":
          description: User order list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

        "400":
          description: Invalid pagination parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{id}:
    parameters:
      - $ref: "#/components/parameters/OrderId"

    get:
      tags:
        - Orders
      summary: Get user order by ID
      description: Returns an order only if it belongs to the authenticated user.
      operationId: getUserOrderById
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Order found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "400":
          description: Invalid order ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid order id

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Order not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: order not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/OrderId"

    patch:
      tags:
        - Orders
      summary: Cancel an order
      description: Cancels an order belonging to the authenticated user when its current status permits cancellation.
      operationId: cancelOrder
      security:
        - BearerAuth: []

      responses:
        "204":
          description: Order successfully cancelled.

        "400":
          description: Invalid order ID or the order cannot be cancelled in its current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Order not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: order not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/orders:
    get:
      tags:
        - Admin
      summary: List all orders
      description: Returns a filtered and paginated order list. Requires the admin role.
      operationId: adminListOrders
      security:
        - BearerAuth: []

      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"

        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1

        - name: from
          in: query
          required: false
          description: Start date in YYYY-MM-DD format using the configured application timezone.
          schema:
            type: string
            format: date
          example: "2026-09-01"

        - name: to
          in: query
          required: false
          description: End date in YYYY-MM-DD format using the configured application timezone. The full day is included.
          schema:
            type: string
            format: date
          example: "2026-09-30"

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

      responses:
        "200":
          description: Filtered order list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"

        "400":
          description: Invalid filter or pagination parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/orders/{id}:
    parameters:
      - $ref: "#/components/parameters/OrderId"

    get:
      tags:
        - Admin
      summary: Get order by ID
      description: Returns any order by ID. Requires the admin role.
      operationId: adminGetOrderById
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Order found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "400":
          description: Invalid order ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: order not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/orders/{id}/status:
    parameters:
      - $ref: "#/components/parameters/OrderId"

    patch:
      tags:
        - Admin
      summary: Update order status
      description: Changes an order status according to the allowed lifecycle transitions. Requires the admin role.
      operationId: adminUpdateOrderStatus
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
            example:
              status: confirmed

      responses:
        "204":
          description: Order status successfully updated.

        "400":
          description: Invalid order ID, request body, status, or status transition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: order not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: id
      in: path
      required: true
      description: Order identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    OrderStatus:
      type: string
      enum:
        - new
        - confirmed
        - cooking
        - ready
        - delivering
        - completed
        - cancelled
      example: new
    CreateOrderItem:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          format: int64
          minimum: 1
          example: 10
        quantity:
          type: integer
          minimum: 1
          example: 2
    CreateOrderRequest:
      type: object
      required:
        - address_id
        - items
      properties:
        address_id:
          type: integer
          format: int64
          minimum: 1
          example: 1
        items:
          type: array
          minItems: 1
          description: Product IDs must be unique within the order.
          items:
            $ref: "#/components/schemas/CreateOrderItem"
    UpdateOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
    OrderItem:
      type: object
      required:
        - id
        - order_id
        - product_id
        - name_snapshot
        - price_snapshot
        - quantity
        - created_at
      properties:
        id:
          type: integer
          format: int64
        order_id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        name_snapshot:
          type: string
        price_snapshot:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    Order:
      type: object
      required:
        - id
        - user_id
        - status
        - total_price
        - delivery_address
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/OrderStatus"
        total_price:
          type: integer
          format: int64
        delivery_address:
          type: string
          description: Delivery address snapshot captured when the order was created.
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderList:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
          format: int64
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0